requirements.txt
absl-py==2.3.0
accelerate==1.8.1
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Deprecated==1.2.18
distro==1.9.0
dotenv==0.9.9
einops==0.8.1
exceptiongroup==1.3.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
future==1.0.0
gast==0.6.0
google-api-core==2.25.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.1
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
keras==2.15.0
kiwisolver==1.4.8
libclang==18.1.1
lightning-utilities==0.14.3
line-bot-sdk==3.17.1
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.3
ml-dtypes==0.2.0
monai==1.5.0
mpmath==1.3.0
multidict==6.6.3
networkx==3.4.2
numpy==1.24.4
oauthlib==3.3.1
openai==1.93.0
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.1
peft==0.16.0
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.10
pydantic==2.11.7
pydantic_core==2.33.2
pyngrok==7.2.11
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
termcolor==3.1.0
threadpoolctl==3.6.0
timm==1.0.16
tokenizers==0.21.2
torch==2.6.0
torchaudio==2.6.0
torchmetrics==1.7.4
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.14.1
yarl==1.20.1
